+++ tools/auto/compileChanged.ts
@@ -2,7 +2,7 @@ import actions from "@actions/core";
 import chalk from "chalk";
 
 import PresenceCompiler from "../classes/PresenceCompiler.js";
-import getDiff from "../util/getDiff.js";
+import { getDiff } from "../util.js";
 
 const compiler = new PresenceCompiler(),
 	changedFolders = getDiff();
