+++ tools/util.ts
@@ -1,0 +1,89 @@
+import { execSync } from "node:child_process";
+import { basename, dirname } from "node:path";
+import { extname } from "node:path";
+
+import actions from "@actions/core";
+import got from "got";
+
+export type ValidEventName = "push" | "pull_request";
+
+export function getDiff(
+	type: "addedModified" | "removed" | "all" = "addedModified"
+): string[] {
+	const commands: Record<ValidEventName, string> = {
+			push: "HEAD HEAD^",
+			pull_request: `origin/main HEAD`,
+		},
+		eventName = process.argv[2] ? validateArg(process.argv[2]) : "pull_request",
+		changedPresenceFolders = execSync(
+			`git --no-pager diff --name-only --diff-filter=${
+				type === "addedModified"
+					? "ACMRTU"
+					: type === "removed"
+					? "D"
+					: "ACMRTUD"
+			} ${commands[eventName]}`
+		)
+			.toString()
+			.split("\n")
+			.filter(file =>
+				["presence.ts", "iframe.ts", "metadata.json"].includes(basename(file))
+			);
+
+	if (!changedPresenceFolders.length) return [];
+
+	return [...new Set(changedPresenceFolders.map(f => basename(dirname(f))))];
+}
+
+function validateArg(arg: string): ValidEventName {
+	if (!["push", "pull_request"].includes(arg))
+		throw new Error(`CI was not called with a valid event name: ${arg}`);
+	return arg as ValidEventName;
+}
+
+export function getFolderLetter(service: string) {
+	const firstLetter = service.at(0)!.toUpperCase();
+
+	if (firstLetter.match(/[A-Z]/)) return firstLetter;
+	if (firstLetter.match(/[0-9]/)) return "0-9";
+	return "#";
+}
+
+export async function getLatestSchema() {
+	const schemas = await got(
+		"https://api.github.com/repos/PreMiD/Schemas/contents/schemas/metadata",
+		{ responseType: "json" }
+	);
+
+	if (schemas.statusCode !== 200 || !Array.isArray(schemas.body)) {
+		actions.setFailed("Could not fetch latest schema");
+		process.exit();
+	}
+
+	const schema = schemas.body.filter(f => f.name.endsWith(".json")).at(-1);
+
+	if (!schema) {
+		actions.setFailed("Could not find latest schema");
+		process.exit();
+	}
+
+	const schemaFile = await got(schema.download_url, { responseType: "json" });
+
+	if (schemaFile.statusCode !== 200 || !schemaFile.body) {
+		actions.setFailed("Could not fetch latest schema file");
+		process.exit();
+	}
+
+	try {
+		return {
+			url: `https://schemas.premid.app/metadata/${schema.name.replace(
+				extname(schema.name),
+				""
+			)}`,
+			schema: schemaFile.body,
+		};
+	} catch (e) {
+		actions.setFailed("Could not parse latest schema");
+		process.exit();
+	}
+}
